requirements.txt
backports-abc==0.5
bleach==2.0.0
decorator==4.0.11
Django==1.11
django-extensions==1.7.8
djangorestframework==3.6.2
entrypoints==0.2.2
html5lib==0.999999999
ipdb==0.10.3
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
typing==3.6.1
wcwidth==0.1.7
webencodings==0.5.1
wheel==0.24.0
widgetsnbextension==2.0.0
